move source picker from core components
