+++ src/packages/core/components/index.ts
@@ -25,7 +25,6 @@ export * from './input-multi-url/index.js';
 export * from './input-number-range/index.js';
 export * from './input-radio-button-list/index.js';
 export * from './input-slider/index.js';
-export * from './input-tree-picker-source/index.js';
 export * from './input-toggle/index.js';
 export * from './input-upload-field/index.js';
 export * from './multiple-color-picker-input/index.js';
