+++ src/packages/property-editors/entry-point.ts
@@ -1,2 +1,1 @@
 import './checkbox-list/components/index.js';
-import './tree-picker/components/index.js';
